Support integer ports in settings
This is honestly more of a change for me to try committing to a project
on GitHub, but it does have some use. This changes allows you to
specify the database port as an integer, e.g. `port: 5432` instead of
forcing a string, e.g. `port: "5432"`. This tripped me up when I
initially setup pfe.